// src/api/models/plantResponseInverterProvider.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Conluz API
 * Conluz is an API-driven application designed for the efficient management of an energy community,enabling the administration of community members and their corresponding supply points and the retrieval of consumption, production data.
 * OpenAPI spec version: 1.0.0
 */

export type PlantResponseInverterProvider = typeof PlantResponseInverterProvider[keyof typeof PlantResponseInverterProvider];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PlantResponseInverterProvider = {
  HUAWEI: 'HUAWEI',
} as const;