// src/api/models/sharingAgreementPartitionCoefficientResponseApplicationState.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Conluz API
 * Conluz is an API-driven application designed for the efficient management of an energy community,enabling the administration of community members and their corresponding supply points and the retrieval of consumption, production data.
 * OpenAPI spec version: 1.0.0
 */

/**
 * Whether the distributor has applied this coefficient yet
 */
export type SharingAgreementPartitionCoefficientResponseApplicationState = typeof SharingAgreementPartitionCoefficientResponseApplicationState[keyof typeof SharingAgreementPartitionCoefficientResponseApplicationState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SharingAgreementPartitionCoefficientResponseApplicationState = {
  PENDING: 'PENDING',
  APPLIED: 'APPLIED',
} as const;